+++ Assets/Scripts/TownGenerator/Buildables/Window.cs
@@ -1,14 +1,8 @@
-using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Unity.VisualScripting;
-using UnityEditor.PackageManager.UI;
 using UnityEngine;
-using UnityEngine.Assertions.Must;
 using UnityEngine.ProBuilder;
-using UnityEngine.ProBuilder.MeshOperations;
 
 public class Window : MonoBehaviour, IBuildable
 {
@@ -36,6 +30,7 @@ public class Window : MonoBehaviour, IBuildable
         outerFrameMesh.transform.SetParent(transform, false);
         Frame outerFrame = outerFrameMesh.AddComponent<Frame>();
         outerFrame.Initialize(CalculateOuterFrame());
+        outerFrame.Data.IsDirty = true;
         return outerFrame;
     }
     private GridFrame CreateInnerFrame()
@@ -45,6 +40,7 @@ public class Window : MonoBehaviour, IBuildable
         innerFrameMesh.transform.SetParent(transform, false);
         GridFrame innerFrame = innerFrameMesh.AddComponent<GridFrame>();
         innerFrame.Initialize(CalculateInnerFrame());
+        innerFrame.Data.IsDirty = true;
         return innerFrame;
     }
     private Pane CreatePane()
@@ -54,6 +50,7 @@ public class Window : MonoBehaviour, IBuildable
         paneMesh.transform.SetParent(transform, false);
         Pane pane = paneMesh.AddComponent<Pane>();
         pane.Initialize(CalculatePane());
+        pane.Data.IsDirty = true;
         return pane;
     }
     private Door CreateLeftShutter(Vector3[] controlPoints)
@@ -63,9 +60,9 @@ public class Window : MonoBehaviour, IBuildable
         leftShutterMesh.transform.SetParent(transform, false);
         Door leftShutter = leftShutterMesh.AddComponent<Door>();
         DoorData data = m_Data.LeftShutter;
-        data.Polygon.ControlPoints = controlPoints;
-        data.Polygon.Normal = m_Data.Polygon.Normal;
+        data.SetPolygon(controlPoints, m_Data.Polygon.Normal);
         leftShutter.Initialize(data);
+        leftShutter.Data.IsDirty = true;
         return leftShutter;
     }
     private Door CreateRightShutter(Vector3[] controlPoints)
@@ -75,9 +72,9 @@ public class Window : MonoBehaviour, IBuildable
         rightShutterMesh.transform.SetParent(transform, false);
         Door rightShutter = rightShutterMesh.AddComponent<Door>();
         DoorData data = m_Data.RightShutter;
-        data.Polygon.ControlPoints = controlPoints;
-        data.Polygon.Normal = m_Data.Polygon.Normal;
+        data.SetPolygon(controlPoints, m_Data.Polygon.Normal);
         rightShutter.Initialize(data);
+        rightShutter.Data.IsDirty = true;
         return rightShutter;
     }
     #endregion
@@ -86,22 +83,26 @@ public class Window : MonoBehaviour, IBuildable
     private FrameData CalculateOuterFrame()
     {
         FrameData frameData = m_Data.OuterFrame;
-        frameData.Polygon.ControlPoints = m_Data.Polygon.ControlPoints;
-        frameData.Polygon.Normal = m_Data.Polygon.Normal;
+        frameData.SetPolygon(m_Data.Polygon.ControlPoints, m_Data.Polygon.Normal);
         return frameData;
     }
     private GridFrameData CalculateInnerFrame()
     {
         GridFrameData frameData = m_Data.InnerFrame;
-        frameData.Polygon.ControlPoints = m_Data.IsOuterFrameActive ? m_OuterFrame.Data.Holes[0].ControlPoints : m_Data.Polygon.ControlPoints;
-        frameData.Polygon.Normal = m_Data.Polygon.Normal;
+        Vector3[] controlPoints = m_Data.IsOuterFrameActive ? m_OuterFrame.Data.Holes[0].ControlPoints : m_Data.Polygon.ControlPoints;
+        Vector3 normal = m_Data.Polygon.Normal;
+        frameData.SetPolygon(controlPoints, normal);
+
         return frameData;
     }
     private Polygon3DData CalculatePane()
     {
         Polygon3DData pane = m_Data.Pane;
-        pane.Polygon.ControlPoints = m_Data.IsOuterFrameActive ? m_OuterFrame.Data.Holes[0].ControlPoints : m_Data.Polygon.ControlPoints;
-        pane.Polygon.Normal = m_Data.Polygon.Normal;
+        Vector3[] controlPoints = m_Data.IsOuterFrameActive ? m_OuterFrame.Data.Holes[0].ControlPoints : m_Data.Polygon.ControlPoints;
+        Vector3 normal = m_Data.Polygon.Normal;
+
+        pane.SetPolygon(controlPoints, normal);
+
         return pane;
     }
     #endregion
@@ -111,12 +112,17 @@ public class Window : MonoBehaviour, IBuildable
     #region Rebuild
     public void Rebuild()
     {
+        if (!m_Data.IsDirty)
+            return;
+
         transform.DeleteChildren();
 
         RebuildOuterFrame();
         RebuildInnerFrame();
         RebuildPane();
         RebuildShutters();
+
+        m_Data.IsDirty = false;
     }
     public void RebuildOuterFrame()
     {
@@ -164,7 +170,7 @@ public class Window : MonoBehaviour, IBuildable
     {
         if (m_Data.IsOuterFrameActive && (m_OuterFrame == null || m_Data.OuterFrame.IsDirty))
         {
-            m_OuterFrame = m_OuterFrame != null ? m_OuterFrame : CreateOuterFrame();
+            m_OuterFrame = m_OuterFrame != null ? m_OuterFrame : CreateOuterFrame(); // What is we need to recalc the points
             m_OuterFrame.Build();
             m_OuterFrame.Data.IsDirty = false;
         }
@@ -221,12 +227,12 @@ public class Window : MonoBehaviour, IBuildable
         if (m_Data.ActiveElements == WindowElement.Nothing)
             return;
 
-        Debug.Log("Window build ", this);
-
         BuildOuterFrame();
         BuildInnerFrame();
         BuildPane();
         BuildShutters();
+
+        Rebuild(); // If window data is dirty rebuild all
 
     }
     #endregion
